Fixed FixedUpdate

+++ src/Sparkle/csharp/Time.cs
@@ -5,7 +5,7 @@ namespace Sparkle.csharp;
 public static class Time {
 
     public static double DeltaTime => Application.Instance.IWindow.Time;
-    public static double TotalTime => (float) _watch.Elapsed.TotalSeconds;
+    public static double TotalTime => _watch.Elapsed.TotalSeconds;
 
     private static Stopwatch _watch;
